+++ inst/modules/exact_p.R
@@ -3,7 +3,9 @@
 #' @description
 #' List any p-values reported with insufficient precision (e.g., p < .05 or p = n.s.)
 #'
-#' @author  Lisa DeBruine (\email{lisa.debruin@glasgow.ac.uk}) and Daniel Lakens (\email{D.Lakens@tue.nl})
+#' @keywords results
+#'
+#' @author  Lisa DeBruine (\email{lisa.debruine@glasgow.ac.uk}) and Daniel Lakens (\email{D.Lakens@tue.nl})
 #'
 #' @import dplyr
 #'
@@ -57,22 +59,36 @@ exact_p <- function(paper, ...) {
     report <- "We detected no *p* values."
     summary_text <- report
   } else if (tl == "green") {
-    report <- "All detected *p* values were reported with appropriate precision."
+    report <- sprintf(
+      "We found no imprecise *p* values out of %d detected.",
+      nrow(p)
+    )
     summary_text <- report
   } else {
     module_output <- sprintf(
-      "We found %d imprecise *p* value%s. Reporting *p* values imprecisely (e.g., *p* < .05) reduces transparency, reproducibility, and re-use (e.g., in *p* value meta-analyses). Best practice is to report exact p-values with three decimal places (e.g., *p* = .032) unless *p* values are smaller than 0.001, in which case you can use *p* < .001.",
-      nrow(report_table), ifelse(nrow(report_table) == 1, "", "s")
+      "We found %d imprecise *p* value%s out of %d detected. Reporting *p* values imprecisely (e.g., *p* < .05) reduces transparency, reproducibility, and re-use (e.g., in *p* value meta-analyses). Best practice is to report exact p-values with three decimal places (e.g., *p* = .032) unless *p* values are smaller than 0.001, in which case you can use *p* < .001.",
+      nrow(report_table), plural(nrow(report_table)), nrow(p)
     )
 
-    summary_text <- sprintf("We found %d imprecise *p* value%s.",
+    summary_text <- sprintf("We found %d imprecise *p* value%s out of %d detected.",
                             nrow(report_table),
-                            ifelse(nrow(report_table) == 1, "", "s"))
+                            plural(nrow(report_table)),
+                            nrow(p))
 
     # Guidance text
+    apa <- bibentry(
+      bibtype = "Book",
+      author  = person("American Psychological Association"),
+      year    = 2020,
+      title   = "Publication manual of the American Psychological Association",
+      edition = "7",
+      subtitle = "The official guide to APA style",
+      publisher = "American Psychological Association"
+    )
+
     guidance <- c(
       "The APA manual states: Report exact *p* values (e.g., *p* = .031) to two or three decimal places. However, report *p* values less than .001 as *p* < .001. However, 2 decimals is too imprecise for many use-cases (e.g., a *p* value meta-analysis), so report *p* values with three digits.",
-      "American Psychological Association. (2020). Publication manual of the American Psychological Association 2020: the official guide to APA style (7th ed.). American Psychological Association."
+      format(apa)
     )
 
     # Combine everything into report text